Enable interface autodiscovery based on a reachable IP
Keepalived-operator assumes that all nodes in a given KeepalivedGroup
will have the same networking configurations and the same interface names.
This is not always true and it makes for a harder setup in non-omogeneus
environments.

This commit introduces an extra optional field on the CRD called
`interfaceFromIP`. If this field is specified, the `interface` field will
be ignored and the interface name will instead be autodiscovered on each
node based on which interface would be used to reach the specified IP.

This commit requires the keepalived-operator image to contain the `ip`, `sed`
and `awk` binaries on the path.

+++ config/docker/notify.sh
@@ -3,16 +3,35 @@
 ## $template contains the template to execute
 ## $verb
 
+function set_up_configs {
+  cp $file $dst_file
+  if [ -n "$IFACE" ]; then
+    sed -i "s/interface.*$/interface $IFACE/g" $dst_file
+    echo "autodicovered local interface that can reach $reachip to be $IFACE"
+  fi
+}
+
 set -o nounset
 set -o errexit
 
 HASH=$(md5sum $(readlink -f $file))
+
+IFACE=""
+if [ -n "$reachip" ]; then
+  IFACE=$(ip route get $reachip | awk "/$reachip/{ print \$3 }")
+fi
+
+if [ "$setup" = "true" ]; then
+  set_up_configs
+  exit 0
+fi
 
 while true; do
    NEW_HASH=$(md5sum $(readlink -f $file))
    if [ "$HASH" != "$NEW_HASH" ]; then
      HASH="$NEW_HASH"
      echo "[$(date +%s)] Trigger refresh"
+     set_up_configs
      kill -SIGHUP $(cat $pid); 
      echo "sent kill signal SIGHUP to $(cat $pid) with outcome $?"
    fi
